Adopt a nice and simple component structure, create separate folders for Presentational and Container Components
